funcionando findCarrersOrderedByStudentCount punto g

+++ src/main/java/com/tudai/integrador3/dto/CareerStudentDto.java
@@ -3,9 +3,9 @@ package com.tudai.integrador3.dto;
 public class CareerStudentDto {
     private int idCareer;
     private String name;
-    private int countStudents;
+    private Long countStudents;
 
-    public CareerStudentDto(int idCareer, String name, int countStudents) {
+    public CareerStudentDto(int idCareer, String name, Long countStudents) {
         this.idCareer = idCareer;
         this.name = name;
         this.countStudents = countStudents;
@@ -19,7 +19,7 @@ public class CareerStudentDto {
         return name;
     }
 
-    public int getCountStudents() {
+    public Long getCountStudents() {
         return countStudents;
     }
 }
